Improve bulk record field error handling
