+++ api/compose/spec.json
@@ -902,7 +902,7 @@
               "title": "Record values"
             },
             {
-              "type": "types.BulkRecordSet",
+              "type": "types.RecordBulkSet",
               "name": "records",
               "required": false,
               "title": "Records"
@@ -948,7 +948,7 @@
               "title": "Record values"
             },
             {
-              "type": "types.BulkRecordSet",
+              "type": "types.RecordBulkSet",
               "name": "records",
               "required": false,
               "title": "Records"
